Changed the topology tools to its own class with methods. The allows methods in the class to call other methods (namely trim and extend)
