+++ OpenPNM/Utilities/__init__.py
@@ -13,5 +13,5 @@ r"""
 from . import IO
 from . import transformations
 from . import misc
-from . import topology
 from . import vertexops
+from .__topology__ import topology
